moved navbar.jsx and refactored import statements
