+++ ourexp-ui/src/components/navbar/Navbar.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { NavLink } from 'react-router-dom'
-import Home from '../pages/Home'
 
 export default function Navbar() {
