+++ ourexp-ui/src/pages/Account.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../components/Navbar'
+import Navbar from '../components/navbar/Navbar'
 
 export default function Account() {
   return (
