+++ ourexp-ui/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../components/Navbar'
+import Navbar from '../components/navbar/Navbar'
 
 export default function Home() {
   return (
