+++ ourexp-ui/src/pages/Submit.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../components/Navbar'
+import Navbar from '../components/navbar/Navbar'
 import SubmissionForm from '../components/submit/SubmissionForm'
 
 const feelingCategories = [
